Limit captured lifetimes of Storage::get and Storage::put
A return-position impl trait (RPIT) captures all lifetimes, but the `path` lifetime should not be captured. The `+ use<...>` syntax cannot be used with stable Rust yet. Thus, this is worked around with associated types and the captures trick.

This should also be done for the `Storage::list`, but there we're using a map with a closure in the `InMemoryStorage` which produces an unnameable types.