Do not capture any lifetime in Storage::get and Storage::put
Only this way, we can implement the necessary logic for the cache. Otherwise, the first `put` would keep a mutable reference preventing to put subsequent files until that reference is dropped, but this is needed because the writing of the meta files needs to be delayed until the main file is closed.

This makes the `InMemoryStorage` also more similar to `FilesystemStorage` in its behavior.